# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a book ordering API built with Golang and Gin framework.",
        "title": "BookStack API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/order": {
            "get": {
                "description": "Get oders by userId",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Get oroders by userId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create an order based on the provided request data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Create a new order",
                "parameters": [
                    {
                        "description": "Order request payload",
                        "name": "order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.OrderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/order/{orderId}/cancel": {
            "post": {
                "description": "Cancel an order by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Cancel an order by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Order ID",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/user": {
            "get": {
                "description": "Retrieve a list of all users",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get all users",
                "responses": {
                    "200": {
                        "description": "Successful retrieval of users",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Service error",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update user details based on the token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "User update request",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UserUpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Update successful",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update user",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        },
        "/user/{userId}": {
            "delete": {
                "description": "Xóa người dùng dựa trên ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Xóa người dùng",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID của người dùng cần xóa",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Xóa người dùng thành công",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "400": {
                        "description": "Không lấy được userId hợp lệ",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    },
                    "500": {
                        "description": "Lỗi server",
                        "schema": {
                            "$ref": "#/definitions/response.WebResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "request.OrderDetailRequest": {
            "type": "object",
            "required": [
                "book_id",
                "quantity"
            ],
            "properties": {
                "book_id": {
                    "description": "ID của sách",
                    "type": "integer"
                },
                "quantity": {
                    "description": "Số lượng sách đặt",
                    "type": "integer"
                }
            }
        },
        "request.OrderRequest": {
            "type": "object",
            "required": [
                "order_details",
                "user_id"
            ],
            "properties": {
                "address": {
                    "description": "Địa chỉ giao hàng",
                    "type": "string"
                },
                "order_details": {
                    "description": "Danh sách sách trong đơn",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.OrderDetailRequest"
                    }
                },
                "phone": {
                    "description": "Số điện thoại",
                    "type": "string"
                },
                "user_id": {
                    "description": "Người đặt hàng",
                    "type": "integer"
                }
            }
        },
...
        "request.UserUpdateRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "response.WebResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}